+++ src/vulkan-srcs/vkUtility.cpp
@@ -298,7 +298,7 @@ namespace vk {
 
 		uint32_t CalculateMipLevels(const uint32_t& imageWidth, const uint32_t& imageHeight) 
 		{
-			return std::floor(std::log2(std::max(imageWidth, imageHeight))) + 1;
+			return static_cast<uint32_t>(std::floor(std::log2(std::max(imageWidth, imageHeight))) + 1);
 
 		}
